refactor(safety): address review findings - sanitizer, tests, structure
- Make sanitizer always run unconditionally (remove injection_check_enabled
  gate in Stage 2). The sanitizer now runs on all content regardless of
  configuration flag.

- Fix proptest string strategy: replace literal pattern with any::<String>()
  to exercise real mixed-width UTF-8 inputs.

- Move test modules to separate files to keep mod.rs under 400 lines:
  * Extract helper_tests to tests/helper_tests.rs
  * Extract main tests to tests/mod_tests.rs
  * Create tests/mod.rs as module entry point
  * mod.rs now 273 lines (was 547)
