+++ src/safety/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Test modules for the safety module.
+
+#[cfg(test)]
+mod helper_tests;
+#[cfg(test)]
+mod mod_tests;
